Add support for duplication with external models
This commit adds a way for using an external action for duplicating the
XML of a block, this allows developers to replace ids in the XML or
execute other side effects of the duplication.

This commit also removes the Scratch specific duplication and reverts to
the original duplication method of Blockly.

Signed-off-by: Juhos Ádám <ajuhos@logiscool.com>